Move legacy config check to step so it can be reused in other pipelines
